refactor: refactor test utils to create a separate kpm home for testing
