+++ pkg/client/client_test.go
@@ -36,25 +36,6 @@ import (
 	"kcl-lang.io/kpm/pkg/utils"
 )
 
-const testDataDir = "test_data"
-
-func getTestDir(subDir string) string {
-	pwd, _ := os.Getwd()
-	testDir := filepath.Join(pwd, testDataDir)
-	testDir = filepath.Join(testDir, subDir)
-
-	return testDir
-}
-
-func initTestDir(subDir string) string {
-	testDir := getTestDir(subDir)
-	// clean the test data
-	_ = os.RemoveAll(testDir)
-	_ = os.Mkdir(testDir, 0755)
-
-	return testDir
-}
-
 func TestWithGlobalLock(t *testing.T) {
 	test.RunTestWithGlobalLock(t, "TestUpdateWithKclMod", testUpdateWithKclMod)
 	test.RunTestWithGlobalLock(t, "TestUpdateWithKclModlock", testUpdateWithKclModlock)
@@ -83,7 +64,6 @@ func TestWithGlobalLock(t *testing.T) {
 	test.RunTestWithGlobalLock(t, "TestDownloadGitWithPackage", testDownloadGitWithPackage)
 	test.RunTestWithGlobalLock(t, "TestModandLockFilesWithGitPackageDownload", testModandLockFilesWithGitPackageDownload)
 	test.RunTestWithGlobalLock(t, "TestDependencyGraph", testDependencyGraph)
-	test.RunTestWithGlobalLock(t, "TestAddWithModSpec", testAddWithModSpec)
 	test.RunTestWithGlobalLock(t, "TestRunRemoteWithArgsInvalid", testRunRemoteWithArgsInvalid)
 	test.RunTestWithGlobalLock(t, "TestRunRemoteWithArgs", testRunRemoteWithArgs)
 	test.RunTestWithGlobalLock(t, "TestRunWithNoSumCheck", testRunWithGitPackage)
@@ -101,7 +81,6 @@ func TestWithGlobalLock(t *testing.T) {
 	test.RunTestWithGlobalLock(t, "TestGraph", testGraph)
 
 	features.Enable(features.SupportNewStorage)
-	test.RunTestWithGlobalLock(t, "testAddWithModSpec", testAddWithModSpec)
 	test.RunTestWithGlobalLock(t, "testAddWithOnlyModSpec", testAddWithOnlyModSpec)
 	test.RunTestWithGlobalLock(t, "testAddRenameWithModSpec", testAddRenameWithModSpec)
 	test.RunTestWithGlobalLock(t, "testAddRenameWithNoSpec", testAddRenameWithNoSpec)
